Moved SelectStyles.js into /pages/people/SelectStyles
